qi: Supports of the non-GNU C++ standard library.
Until now, we only supported the GNU C++ standard library. On more recent
systems, the extensions were colliding with our use of macros in
Qi Log, and only there.

Now, using Boost VMD (which comes with Boost 1.60), it is possible to
write the problematic macros in a way that is supported by both libraries.

It enables dependent projects to use safely libQi jointly with other
libraries (like ROS) that require the non-GNU standard library.

Change-Id: I1ee91bc8ceb3f7b5cc23936a5daa332f1d6656f3